[change] Merge two commit methods
